Rename decorators module to debug_metadata
Removes the pre-#3192 vocabulary from core/src/operations. The module
contains only debug/source metadata types (AssemblyOp, DebugVarInfo,
DebugVarLocation), not runtime decorators.

Part of the post-#3192 simplification plan (Pass 1).
