+++ core/src/operations/debug_metadata/assembly_op.rs
