+++ core/src/operations/debug_metadata/mod.rs
