+++ core/src/operations/mod.rs
@@ -3,8 +3,8 @@ use core::fmt;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
-mod decorators;
-pub use decorators::{AssemblyOp, DebugVarInfo, DebugVarLocation};
+mod debug_metadata;
+pub use debug_metadata::{AssemblyOp, DebugVarInfo, DebugVarLocation};
 
 use crate::{
     Felt,
